pyproject.toml
[tool.poetry]
name = "aerleon"
version = "1.0.1"
description = "A firewall generation tool"
authors = ["Rob Ankeny <ankenyr@gmail.com>", "Jason Benterou <jason.benterou@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/aerleon/aerleon"
repository = "https://github.com/aerleon/aerleon"
keywords = ["firewall", "networking", "security"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Security",
    "Topic :: System :: Networking :: Firewalls"
]
include = [
    "LICENSE",
    "README.md",
]
packages = [{include = "aerleon"}, {include = "aerleon/lib"}]

[tool.poetry.dependencies]
python = "^3.7"
PyYAML = "^6.0"
ply = "^3.11"
absl-py = "^1.2.0"
typing_extensions = "^4.4.0"
importlib-metadata = { version = "^4.2", markers = "python_version <= '3.10'" }

[tool.poetry.dev-dependencies]
black = {version = "^22.8.0"}
coverage = {version = "^6.4.4"}
flake8 = {version = "*"}
Flake8-pyproject = {version = "*"}
nox = {version = "^2022.8.7"}
nox-poetry = {version = "^1.0.1"}
pre-commit = {version = "^2.20.0"}
pytest = {version = "^7.1.3"}
pyperf = {version = "^2.4.1"}
psutil = {version = "^5.9.2"}
mkdocs = "1.3.1"
mkdocs-material = "8.3.9"
mkdocstrings = "0.19"
mkdocstrings-python = "0.7.1"
mkdocs-version-annotations = "1.0.0"
sigstore = "^1.0.0"

[tool.poetry.scripts]
aclgen = "aerleon.aclgen:EntryPoint"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
# extend-exclude = "/__pycache__|\.ref/"
line-length = 99
skip-string-normalization = true
target-version = ['py37']
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | settings.py     # This is where you define files that should not be stylized by black
                     # the root of the project
)
'''

[tool.isort]
profile = "black"

[tool.flake8]
max-complexity = 10
max-line-length = 99
extend-ignore = ['E203', 'C901']
exclude = ['.git','.github','venv','site-packages','__pycache__','doc','build','dist','policies']

[tool.coverage.run]
branch = true
command_line = '-m pytest'
